Move camera code to new folder
