+++ libs/vxng/CMakeLists.txt
@@ -3,8 +3,8 @@ project(vxng)
 add_library(${PROJECT_NAME}
     src/glsl/fullscreen.frag.cpp
     src/glsl/fullscreen.vert.cpp
-    src/scene/camera.cpp
-    src/scene/orbit-camera.cpp
+    src/camera/camera.cpp
+    src/camera/orbit-camera.cpp
     src/scene/scene.cpp
     src/gl-util.cpp
     src/renderer.cpp
