+++ libs/vxng/include/vxng/camera.h
