+++ libs/vxng/include/vxng/orbit-camera.h
